% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_DES.R
\name{boot_DES}
\alias{boot_DES}
\title{Bootstrap fossil biogeography}
\usage{
boot_DES(Nsim = 10, Time, Step, BinSize, Nspecies, SimD, SimE, SimQ,
  Qtimes = NULL, Origin = "random", VarD = NULL, VarE = NULL,
  Covariate = NULL, DivD = NULL, DivE = NULL, DdE = NULL,
  Cor = "exponential", DataInArea = NULL, Ncat = NULL,
  alpha = NULL, Observation = NULL, ConfInt = 0.95, Ncores = 1)
}
\arguments{
\item{Nsim}{Number of bootstrap simulations}

\item{Time}{Positive number indicating the time covered by the simulation}

\item{Step}{Time steps for the simulation}

\item{BinSize}{Size of the time bins for the simulated data}

\item{Nspecies}{Number of species}

\item{SimD}{Dispersal rates}

\item{SimE}{Extinction rates}

\item{SimQ}{Sampling rates}

\item{Qtimes}{Shift times in dispersal, extinction, and sampling}

\item{Origin}{Area of origin \cr Options 'random', '1', or '2'}

\item{VarD}{Strengths of covariate dependent dispersal
and in case of logistic correlation the midpoints}

\item{VarE}{Strengths of covariate dependent extinction
and in case of logistic correlation the midpoints}

\item{Covariate}{data.frame with the covariate for
e.g. environmental dependent dispersal or extinction}

\item{DivD}{Strengths of diversity dependent dispersal}

\item{DivE}{Strengths of diversity dependent extinction}

\item{DdE}{Strengths of dispersal dependent extinction}

\item{Cor}{Correlation with environment or diversity \cr Options 'exponential' or 'logistic'}

\item{DataInArea}{Simulate dynamic in that area}

\item{Ncat}{Number of categories for the discretized Gamma distribution
to simulate heterogeneous sampling}

\item{alpha}{i.e. the shape and rate of the Gamma distribution}

\item{Observation}{Two column matrix or data.frame of first observation time and area}

\item{ConfInt}{Confidence interval for simualted diversities
(e.g. c(0.95, 0.68) for two and one standard deviation)}

\item{Ncores}{Number of CPU cores to use for simulations}
}
\value{
A list of three elements.
}
\description{
This function simulates several times the biogeographic events
dispersal, extinction, and sampling to obtain diversity trajectories
and their associated uncertainties through time
}
\author{
Torsten Hauffe
}
