% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsm_DES.R
\name{bsm_DES}
\alias{bsm_DES}
\title{Biogeographic stochastic mapping for fossil biogeography}
\usage{
bsm_DES(
  DesIn,
  NumBsm = 1,
  MaxAttempts = 1000,
  Step,
  SimD,
  SimE,
  Qtimes = NULL,
  VarD = NULL,
  VarE = NULL,
  Covariate = NULL,
  DivD = NULL,
  DivE = NULL,
  DdE = NULL,
  Cor = "exponential",
  DataInArea = NULL,
  TraitD = NULL,
  VarTraitD = NULL,
  TraitE = NULL,
  VarTraitE = NULL,
  Verbose = FALSE
)
}
\arguments{
\item{DesIn}{Input file as for a PyRateDES2 analysis}

\item{NumBsm}{Number of biogeographic stochastic maps}

\item{MaxAttempts}{Maximum number of attempts to stochastically simulate the biogeographic history of a single taxon}

\item{Step}{Time steps for the simulation}

\item{SimD}{Dispersal rates}

\item{SimE}{Extinction rates}

\item{Qtimes}{Shift times in dispersal, extinction, and sampling}

\item{VarD}{Strengths of covariate dependent dispersal
and in case of logistic correlation the midpoints}

\item{VarE}{Strengths of covariate dependent extinction
and in case of logistic correlation the midpoints}

\item{Covariate}{data.frame with the covariate for
e.g. environmental dependent dispersal or extinction}

\item{DivD}{Strengths of diversity dependent dispersal}

\item{DivE}{Strengths of diversity dependent extinction}

\item{DdE}{Strengths of dispersal dependent extinction}

\item{Cor}{Correlation with environment or diversity \cr Options 'exponential' or 'logistic'}

\item{DataInArea}{Simulate dynamic in that area}

\item{TraitD}{Data.frame or matrix of a trait influencing dispersal (first column taxon index and second the trait). Either a continuous trait or a binary trait codes as 1 and exp(1) (internal log transformation)}

\item{VarTraitD}{Strengths of trait dependent dispersal}

\item{TraitE}{Continuous influencing extinction (first column taxon index and second the trait). See TraitD for details}

\item{VarTraitE}{Strengths of trait dependent extinction}

\item{Verbose}{Should progress be printed?}
}
\value{
A list with stochastically mapped biogeographic histories
}
\description{
This function performs a stochastically simulates the biogeographic history of taxa
}
\author{
Torsten Hauffe
}
