% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim_df.R
\name{gen_sim_df}
\alias{gen_sim_df}
\title{Template to simulate fossil biogeography}
\usage{
gen_sim_df(Time, Step, Nspecies, Origin = "random", Covariate = NULL)
}
\arguments{
\item{Time}{Positive number indicating the}

\item{Step}{Time steps for the simulation.}

\item{Nspecies}{Number of species}

\item{Origin}{Area of origin \cr Options '1', '2' or '3'}

\item{Covariate}{Covariate for e.g. environmental dependent dispersal or extinction}
}
\value{
A data.frame
}
\description{
This function generates a data.frame
to simulate dispersal and extinction through time.
}
\author{
Torsten Hauffe
}
