% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gamma_rates.R
\name{get_gamma_rates}
\alias{get_gamma_rates}
\title{Discretized Gamma distribution}
\usage{
get_gamma_rates(alpha, Ncat)
}
\arguments{
\item{alpha}{i.e. the shape and rate of the Gamma distribution}

\item{Ncat}{Number of categories for the discretized Gamma distribution}
}
\value{
Discretized gamma distribution
}
\description{
This function discretizes the Gamma distribution and
is used to simulate heterogeneous fossil preservation via multiplication
with the preservation rate
}
\references{
Yang, Z. (1993).
Maximum-likelihood estimation of phylogeny from DNA sequences
when substitution rates differ over sites.
Molecular Biology and Evolution 10(6): 1396-1401.
}
\author{
Torsten Hauffe
}
