% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_num_dis_ext_events.R
\name{get_num_dis_ext_events}
\alias{get_num_dis_ext_events}
\title{Number of dispersal and extinction events}
\usage{
get_num_dis_ext_events(Bsm)
}
\arguments{
\item{Bsm}{The output of bsm_DES: A matrix (in a case of a single biogeographic stochastic map) or a list of matrices (in a case of several biogeographic stochastic maps)}
}
\value{
A list of two data.frames with the number of dispersal and extinction events per stochastic mapping
}
\description{
This function calculates for each taxon the number of dispersal and extinction events of a stochastic mapping of its biogeographic history
}
\author{
Torsten Hauffe
}
