% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boot_DES.R
\name{plot_boot_DES}
\alias{plot_boot_DES}
\title{Plot bootstrapped fossil biogeography}
\usage{
plot_boot_DES(
  BootDes,
  DivToPlot = 1:3,
  Col = c("orange", "dodgerblue", "green"),
  Transparency = 0.5
)
}
\arguments{
\item{BootDes}{Bootstrapped fossil biogeography from boot_DES}

\item{DivToPlot}{Diversities to plot}

\item{Col}{Colors for the different diversities}

\item{Transparency}{Transparency for the confidence intervals}
}
\value{
A plot
}
\description{
This function plots diversity trajectories
and their associated uncertainties through time
}
\author{
Torsten Hauffe
}
