% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim_df.R
\name{gen_sim_df}
\alias{gen_sim_df}
\title{Template to simulate fossil biogeography}
\usage{
gen_sim_df(TimeSim, Nspecies, Origin = "random", Qtimes = NULL,
  Covariate = NULL, DataInArea = NULL, Ncat = NULL,
  Observation = NULL, GlobExt = NULL)
}
\arguments{
\item{TimeSim}{Time steps for the simulation}

\item{Nspecies}{Number of species}

\item{Origin}{Area of origin \cr Options 'random', '1', or '2'}

\item{Qtimes}{Shift times in dispersal, extinction, and sampling}

\item{Covariate}{data.frame with the covariate for
e.g. environmental dependent dispersal or extinction}

\item{DataInArea}{Simulate dynamic in that area}

\item{Ncat}{Number of categories for the discretized Gamma distribution
to simulate heterogeneous sampling}

\item{Observation}{Two column matrix or data.frame of first observation time and area}

\item{GlobExt}{Global extinction rate}

\item{alpha}{i.e. the shape and rate of the Gamma distribution}
}
\value{
A data.frame
}
\description{
This function generates a data.frame
to simulate dispersal and extinction through time.
}
\author{
Torsten Hauffe
}
