% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_DES.R
\name{sim_DES}
\alias{sim_DES}
\title{Simulate fossil biogeography}
\usage{
sim_DES(
  Time,
  Step,
  BinSize,
  Nspecies,
  SimD,
  SimE,
  SimQ,
  Qtimes = NULL,
  Origin = 0,
  VarD = NULL,
  VarE = NULL,
  Covariate = NULL,
  DivD = NULL,
  DivE = NULL,
  DdE = NULL,
  Cor = "exponential",
  DataInArea = NULL,
  Ncat = NULL,
  alpha = NULL,
  Observation = NULL,
  GlobExt = NULL
)
}
\arguments{
\item{Time}{Positive number indicating the time covered by the simulation}

\item{Step}{Time steps for the simulation}

\item{BinSize}{Size of the time bins for the simulated data}

\item{Nspecies}{Number of species}

\item{SimD}{Dispersal rates}

\item{SimE}{Extinction rates}

\item{SimQ}{Sampling rates}

\item{Qtimes}{Shift times in dispersal, extinction, and sampling}

\item{Origin}{Vector with the possible areas of origin \cr Options: 0 (i.e. random, default) or any combination of 1, 2, and/or 3}

\item{VarD}{Strengths of covariate dependent dispersal
and in case of logistic correlation the midpoints}

\item{VarE}{Strengths of covariate dependent extinction
and in case of logistic correlation the midpoints}

\item{Covariate}{data.frame with the covariate for
e.g. environmental dependent dispersal or extinction}

\item{DivD}{Strengths of diversity dependent dispersal}

\item{DivE}{Strengths of diversity dependent extinction}

\item{DdE}{Strengths of dispersal dependent extinction}

\item{Cor}{Correlation with environment or diversity \cr Options 'exponential' or 'logistic'}

\item{DataInArea}{Simulate dynamic in that area}

\item{Ncat}{Number of categories for the discretized Gamma distribution
to simulate heterogeneous sampling or Inf for a random draw of the Gamma distribution}

\item{alpha}{i.e. the shape and rate of the Gamma distribution}

\item{Observation}{Two column matrix or data.frame of first observation time and area}

\item{GlobExt}{Global extinction rate}
}
\value{
A list with four elements
}
\description{
This function simulates dispersal, extinction, and sampling through time
}
\author{
Torsten Hauffe
}
